requirements.txt
faiss-cpu>=1.8.0
langchain-text-splitters>=0.3.0
numpy>=1.26.0
openai>=1.51.0
python-dotenv>=1.0.1
pypdf>=5.0.0

